# Backend/Frontend/openapi.raw.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/tickets": {
      "post": {
        "operationId": "TicketsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTicketDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Insufficient role permissions"
          },
          "404": {
            "description": "Category not found (or inactive)"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a new ticket",
        "tags": [
          "tickets"
        ]
      },
      "get": {
        "operationId": "TicketsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/TicketStatus"
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/TicketPriority"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "assigneeId",
            "required": false,
            "in": "query",
            "description": "Honored for ADMIN callers only",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "createdById",
            "required": false,
            "in": "query",
            "description": "Honored for ADMIN callers only",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Free-text search on title and reference (case-insensitive)",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "createdAt",
              "type": "string",
              "enum": [
                "createdAt",
                "priority",
                "slaDueAt",
                "status"
              ]
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "default": "DESC",
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/TicketListItemDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List tickets, paginated and filtered, scoped by the caller's role",
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/tickets/{id}": {
      "get": {
        "operationId": "TicketsController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Not the owner, the assignee, nor an admin"
          },
          "404": {
            "description": "Ticket not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get a single ticket by id",
        "tags": [
          "tickets"
        ]
      },
      "patch": {
        "operationId": "TicketsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTicketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "No field provided to update"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Not an admin, not the owning CLIENT with the ticket still OPEN (nor the owner at all), or a TECHNICIAN"
          },
          "404": {
            "description": "Ticket not found, or the given categoryId is unknown/inactive"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update mutable fields of a ticket (role x status rules enforced by the service)",
        "tags": [
          "tickets"
        ]
      },
      "delete": {
        "operationId": "TicketsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Ticket soft deleted"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Caller is not an ADMIN\n\nInsufficient role permissions"
          },
          "404": {
            "description": "Ticket not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Soft delete a ticket (ADMIN only)",
        "tags": [
          "tickets"
        ]
      }
    },
...
    "/api/tickets/{id}/comments": {
      "post": {
        "operationId": "TicketCommentsController_create",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "body missing/empty/too long, or an invalid visibility value"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "OwnershipGuard rejected the caller (not owner/assignee/admin), or a CLIENT attempted to post an INTERNAL comment"
          },
          "404": {
            "description": "Ticket not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Post a comment on a ticket",
        "tags": [
          "tickets"
        ]
      },
      "get": {
        "operationId": "TicketCommentsController_list",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/CommentResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Not the owner, the assignee, nor an admin"
          },
          "404": {
            "description": "Ticket not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List a ticket's comments, paginated and sorted chronologically (createdAt ASC); INTERNAL comments are excluded for a CLIENT caller",
        "tags": [
          "tickets"
        ]
      }
    },
...
  },
  "info": {
    "title": "Ticket Checker API",
    "description": "",
    "version": "1.0",
    "contact": {

    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "example": 42,
            "description": "Total number of items across all pages"
          },
          "page": {
            "type": "number",
            "example": 1,
            "description": "Current, 1-based page number"
          },
          "limit": {
            "type": "number",
            "example": 20,
            "description": "Number of items per page"
          },
          "totalPages": {
            "type": "number",
            "example": 3,
            "description": "Total number of pages"
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
      "PaginatedResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "Page of results. The concrete item type is documented per-endpoint by @ApiPaginatedResponse()."
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "TicketPriority": {
        "type": "string",
        "enum": [
          "LOW",
          "NORMAL",
          "HIGH",
          "CRITICAL"
        ]
      },
      "CreateTicketDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Panne de climatisation salle serveur",
            "minLength": 3,
            "maxLength": 150
          },
          "description": {
            "type": "string",
            "example": "La climatisation ne démarre plus depuis ce matin.",
            "minLength": 1,
            "maxLength": 5000
          },
          "priority": {
            "default": "NORMAL",
            "allOf": [
              {
                "$ref": "#/components/schemas/TicketPriority"
              }
            ]
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "siteLabel": {
            "type": "string",
            "maxLength": 150
          },
          "siteAddress": {
            "type": "string",
            "maxLength": 2000
          }
        },
        "required": [
          "title",
          "description",
          "categoryId"
        ]
      },
      "TicketStatus": {
        "type": "string",
        "enum": [
          "OPEN",
          "ASSIGNED",
          "IN_PROGRESS",
          "RESOLVED",
          "CLOSED",
          "CANCELLED"
        ]
      },
      "TicketCategorySummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "UserSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "username",
          "firstName",
          "lastName"
        ]
      },
      "TicketResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "reference": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TicketStatus"
              }
            ]
          },
          "priority": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TicketPriority"
              }
            ]
          },
          "category": {
            "$ref": "#/components/schemas/TicketCategorySummaryDto"
          },
          "createdBy": {
            "$ref": "#/components/schemas/UserSummaryDto"
          },
          "assignee": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserSummaryDto"
              }
            ]
          },
          "siteLabel": {
            "type": "string",
            "nullable": true
          },
          "siteAddress": {
            "type": "string",
            "nullable": true
          },
          "slaDueAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "assignedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "startedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "resolvedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "closedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "cancelledAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "resolutionNote": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "reference",
          "title",
          "description",
          "status",
          "priority",
          "category",
          "createdBy",
          "assignee",
          "siteLabel",
          "siteAddress",
          "slaDueAt",
          "assignedAt",
          "startedAt",
          "resolvedAt",
          "closedAt",
          "cancelledAt",
          "resolutionNote",
          "createdAt",
          "updatedAt"
        ]
      },
      "TicketListCategorySummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "TicketListAssigneeSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "username"
        ]
      },
      "TicketListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "reference": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TicketStatus"
              }
            ]
          },
          "priority": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TicketPriority"
              }
            ]
          },
          "category": {
            "$ref": "#/components/schemas/TicketListCategorySummaryDto"
          },
          "assignee": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/TicketListAssigneeSummaryDto"
              }
            ]
          },
          "slaDueAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "reference",
          "title",
          "status",
          "priority",
          "category",
          "assignee",
          "slaDueAt",
          "createdAt"
        ]
      },
      "UpdateTicketDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Panne de climatisation salle serveur",
            "minLength": 3,
            "maxLength": 150
          },
          "description": {
            "type": "string",
            "example": "La climatisation ne démarre plus depuis ce matin.",
            "minLength": 1,
            "maxLength": 5000
          },
          "priority": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TicketPriority"
              }
            ]
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "siteLabel": {
            "type": "string",
            "maxLength": 150
          },
          "siteAddress": {
            "type": "string",
            "maxLength": 2000
          }
        }
      },
...
      "CommentVisibility": {
        "type": "string",
        "enum": [
          "PUBLIC",
          "INTERNAL"
        ],
        "description": "INTERNAL is reserved to ADMIN/TECHNICIAN — a CLIENT sending INTERNAL is rejected with 403 (never silently downgraded to PUBLIC)."
      },
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "example": "Le technicien est en route, arrivée estimée dans 30 minutes.",
            "minLength": 1,
            "maxLength": 5000
          },
          "visibility": {
            "default": "PUBLIC",
            "description": "INTERNAL is reserved to ADMIN/TECHNICIAN — a CLIENT sending INTERNAL is rejected with 403 (never silently downgraded to PUBLIC).",
            "allOf": [
              {
                "$ref": "#/components/schemas/CommentVisibility"
              }
            ]
          }
        },
        "required": [
          "body"
        ]
      },
      "CommentAuthorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "username"
        ]
      },
      "CommentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "visibility": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CommentVisibility"
              }
            ]
          },
          "author": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/CommentAuthorDto"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "body",
          "visibility",
          "author",
          "createdAt"
        ]
      },
...
    }
  }
}